refactor: rename feedback.error file and some constants
