+++ src/app/modules/feedback/feedback.service.ts
@@ -13,7 +13,7 @@ import getFormFeedbackModel from '../../models/form_feedback.server.model'
 import { getMongoErrorMessage } from '../../utils/handle-mongo-error'
 import { DatabaseError } from '../core/core.errors'
 
-import { DuplicateFeedbackSubmissionError } from './feedback.error'
+import { DuplicateFeedbackSubmissionError } from './feedback.errors'
 
 const FormFeedbackModel = getFormFeedbackModel(mongoose)
 const logger = createLoggerWithLabel(module)
