+++ src/app/modules/submission/submission.service.ts
@@ -14,7 +14,7 @@ import { AutoReplyMailData } from '../../services/mail/mail.types'
 import { createQueryWithDateParam, isMalformedDate } from '../../utils/date'
 import { getMongoErrorMessage } from '../../utils/handle-mongo-error'
 import { DatabaseError, MalformedParametersError } from '../core/core.errors'
-import { InvalidSubmissionIdError } from '../feedback/feedback.error'
+import { InvalidSubmissionIdError } from '../feedback/feedback.errors'
 
 import { SendEmailConfirmationError } from './submission.errors'
